backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
numpy==1.24.3
gtts==2.4.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
stripe==7.4.0
sendgrid==6.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
bleach==6.1.0
pydantic[email]==2.5.0
httpx==0.25.2
